+++ roadgraph_builder/routing/__init__.py
@@ -1,0 +1,5 @@
+"""Routing helpers on ``Graph`` (shortest path, etc)."""
+
+from roadgraph_builder.routing.shortest_path import Route, shortest_path
+
+__all__ = ["Route", "shortest_path"]